+++ src/index.ts
@@ -2,5 +2,3 @@ export { Client, EcstarClientOptions } from './Client';
 
 export { command, commandOptions } from './lib/structures/command';
 export { event, eventOptions } from './lib/structures/event';
-
-export { context } from './lib/context';
