+++ src/lib/context/slashCommand.ts
@@ -1,6 +1,6 @@
 import { Interaction, Message, Snowflake, TextChannel } from 'discord.js';
 import { Client } from 'ecstar';
-import { ContextBase } from 'ecstar/context';
+import { ContextBase } from 'ecstar/context/base';
 
 export interface SlashCommandContext extends ContextBase {
   type: 'slashCommand';
